Introduce CurrencyExchange

+++ lib/ad_parameters/creative.rb
@@ -5,7 +5,7 @@ module AdParameters
     extend Buildable
 
     attribute :id, Types::Coercible::String
-    attribute :price, Types::Coercible::Float
+    attribute :price, Types::Coercible::Decimal
     attribute :currency, Types::Coercible::String
     # TODO: Handle currency calculation
     # TODO: Handle comparison with currency
